# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shinda 24/7 Rocket Crash Game API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type GamePhase = (typeof GamePhase)[keyof typeof GamePhase];

export const GamePhase = {
  waiting: "waiting",
  flying: "flying",
  crashed: "crashed",
} as const;

export interface GameState {
  phase: GamePhase;
  multiplier: number;
  crashedAt?: number | null;
  roundId: number;
  countdownMs?: number | null;
  onlineCount: number;
  playingCount: number;
}

export interface GameRound {
  id: number;
  crashedAt: number;
  hash: string;
  createdAt: string;
}

export interface PlaceBetRequest {
  amount: number;
  autoCashOut?: number | null;
  playerId: string;
}

export interface PlaceBetResponse {
  success: boolean;
  betId: number;
  balance: number;
}

export interface CashOutRequest {
  betId: number;
  playerId: string;
}

export interface CashOutResponse {
  success: boolean;
  profit: number;
  multiplier: number;
  balance: number;
}

export interface ActivePlayer {
  id: string;
  username: string;
  amount: number;
  multiplier?: number | null;
  profit?: number | null;
  cashedOut: boolean;
  hash: string;
}

export interface LeaderboardPlayer {
  username: string;
  multiplier: number;
  amount: number;
}

export interface WalletInfo {
  balance: number;
  playerId: string;
}

export interface RegisterRequest {
  phone: string;
  username: string;
}

export interface LoginRequest {
  phone: string;
}

export interface VerifyOtpRequest {
  phone: string;
  otp: string;
  username?: string;
}

export interface OtpSentResponse {
  message: string;
  phone?: string;
}

export interface UserProfile {
  id: number;
  phone: string;
  username: string;
  isAdmin: boolean;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface DepositRequest {
  phone: string;
  amountKes: number;
}

export interface DepositResponse {
  message: string;
  checkoutRequestId?: string;
}

export interface WithdrawRequest {
  phone: string;
  amountKes: number;
}

export interface WithdrawResponse {
  message: string;
  conversationId?: string;
  newBalanceCents?: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetGameHistoryParams = {
  limit?: number;
};