# lib/api-zod/src/generated/types/depositResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shinda 24/7 Rocket Crash Game API
 * OpenAPI spec version: 0.2.0
 */

export interface DepositResponse {
  message: string;
  checkoutRequestId?: string;
}