# lib/api-zod/src/generated/types/gameState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shinda 24/7 Rocket Crash Game API
 * OpenAPI spec version: 0.2.0
 */
import type { GamePhase } from "./gamePhase";

export interface GameState {
  phase: GamePhase;
  multiplier: number;
  crashedAt?: number | null;
  roundId: number;
  countdownMs?: number | null;
  onlineCount: number;
  playingCount: number;
}